[KGB-166]OfficerShiftList is a paginated table
